// apps/web/lib/types/system.ts
export interface StorageResourceSettings {
  enabled: boolean;
}

export interface StorageGoCacheSettings {
  enabled: boolean;
  max_bytes: number;
  adopted_path: string;
}

export interface StorageDockerSettings {
  dedicated_daemon_acknowledged: boolean;
  build_cache_enabled: boolean;
  build_cache_keep_bytes: number;
  build_cache_unused_hours: number;
  unused_images_enabled: boolean;
  unused_images_hours: number;
}

export interface StorageMaintenanceSettings {
  enabled: boolean;
  check_interval_hours: number;
  idle_for_minutes: number;
  orphan_grace_hours: number;
  quarantine_retention_hours: number;
  workspaces: StorageResourceSettings;
  kandev_containers: StorageResourceSettings;
  go_cache: StorageGoCacheSettings;
  docker: StorageDockerSettings;
}

export interface StorageCapabilities {
  managed_go_cache_path: string;
  go_cache_adoption_available: boolean;
  docker_available: boolean;
  docker_host: string;
  host_global_docker_cleanup_allowed: boolean;
}